- Most features are worked

+++ src/Acme.BookStore.Application/Authors/AuthorAppService.cs
@@ -1,35 +1,35 @@
-﻿using Acme.BookStore.Emails;
+﻿using Acme.BookStore.Books;
+using Acme.BookStore.Notifications;
 using Acme.BookStore.Permissions;
+using AutoFilterer.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Dynamic.Core;
 using System.Threading.Tasks;
+using Volo.Abp;
 using Volo.Abp.Application.Dtos;
-using Volo.Abp.BackgroundJobs;
+using Volo.Abp.Data;
 using Volo.Abp.Domain.Repositories;
 using Volo.Abp.Users;
 
 namespace Acme.BookStore.Authors;
 
 [Authorize(BookStorePermissions.Authors.Default)]
-public class AuthorAppService : BookStoreAppService, IAuthorAppService
+public class AuthorAppService(
+    IAuthorRepository authorRepository,
+    AuthorManager authorManager,
+    ICurrentUser currentUser,
+    INotificationAppService notificationAppService,
+    IRepository<Book, Guid> bookRepository
+    ) : BookStoreAppService, IAuthorAppService
 {
-    private readonly IAuthorRepository _authorRepository;
-    private readonly AuthorManager _authorManager;
-    private readonly IBackgroundJobManager _backgroundJobManager;
-    private readonly ICurrentUser _currentUser;
-    public AuthorAppService(
-        IAuthorRepository authorRepository,
-        AuthorManager authorManager,
-        IBackgroundJobManager backgroundJobManager,
-        ICurrentUser currentUser
-        )
-    {
-        _authorRepository = authorRepository;
-        _authorManager = authorManager;
-        _backgroundJobManager = backgroundJobManager;
-        _currentUser = currentUser;
-    }
+    private readonly IAuthorRepository _authorRepository = authorRepository;
+    private readonly AuthorManager _authorManager = authorManager;
+    private readonly ICurrentUser _currentUser = currentUser;
+    private readonly INotificationAppService _notificationAppService = notificationAppService;
+    private readonly IRepository<Book, Guid> _bookRepository = bookRepository;
 
     public async Task<AuthorDto> GetAsync(Guid id)
     {
@@ -39,26 +39,19 @@ public class AuthorAppService : BookStoreAppService, IAuthorAppService
 
     public async Task<PagedResultDto<AuthorDto>> GetListAsync(GetAuthorListDto input)
     {
-        if (input.Sorting.IsNullOrWhiteSpace())
-        {
-            input.Sorting = nameof(Author.Name);
-        }
+        var authors = await _authorRepository.GetQueryableAsync();
 
-        var authors = await _authorRepository.GetListAsync(
-            input.SkipCount,
-            input.MaxResultCount,
-            input.Sorting,
-            input.Filter
-        );
+        authors = authors.ApplyFilter(input);
 
-        var totalCount = input.Filter == null
-            ? await _authorRepository.CountAsync()
-            : await _authorRepository.CountAsync(
-                author => author.Name.Contains(input.Filter));
+        var totalCount = await AsyncExecuter.CountAsync(authors);
+
+        authors = authors.OrderBy(NormalizeSorting(input.Sorting)).PageBy(input.SkipCount, input.MaxResultCount);
+
+        var items = await AsyncExecuter.ToListAsync(authors);
 
         return new PagedResultDto<AuthorDto>(
             totalCount,
-            ObjectMapper.Map<List<Author>, List<AuthorDto>>(authors)
+            ObjectMapper.Map<List<Author>, List<AuthorDto>>(items)
         );
     }
 
@@ -71,12 +64,14 @@ public class AuthorAppService : BookStoreAppService, IAuthorAppService
         );
 
         await _authorRepository.InsertAsync(author);
-        await _backgroundJobManager.EnqueueAsync(
-            new EmailSendingArgs(
-                "phucanhbtt@gmail.com",
-                $"An author has been created by user {_currentUser.Name + _currentUser.SurName}.",
-                $"Author {author.Name} has just been created.")
-        );
+
+        await _notificationAppService.InsertNotificationAndSendEmailAsync(
+            _currentUser.Id.GetValueOrDefault(),
+            _currentUser.Email,
+            NotificationType.AuthorCRUD,
+            ObjectMapper.Map<CreateAuthorDto, AuthorDto>(input),
+            "created");
+
         return ObjectMapper.Map<Author, AuthorDto>(author);
     }
 
@@ -94,27 +89,71 @@ public class AuthorAppService : BookStoreAppService, IAuthorAppService
 
         await _authorRepository.UpdateAsync(author);
 
-        await _backgroundJobManager.EnqueueAsync(
-            new EmailSendingArgs(
-                "phucanhbtt@gmail.com",
-                $"An author has been updated by user {_currentUser.Name + _currentUser.SurName}.",
-                $"Author {author.Name} has just been updated."
-                )
-            );
+        await _notificationAppService.InsertNotificationAndSendEmailAsync(
+            _currentUser.Id.GetValueOrDefault(),
+            _currentUser.Email,
+            NotificationType.AuthorCRUD,
+            ObjectMapper.Map<Author, AuthorDto>(author),
+            "updated");
     }
 
     [Authorize(BookStorePermissions.Authors.Delete)]
     public async Task DeleteAsync(Guid id)
     {
         var author = await _authorRepository.GetAsync(id);
+        // Check if there is any referential entity associated with this lookup entity to ensure referential integrity constraint
+        if (await _bookRepository.AnyAsync(x => x.AuthorId == author.Id))
+        {
+            throw new UserFriendlyException(
+                $"Author '{author.Name}' is associated with existing books.",
+                "405",
+                $"Please delete or reassign the books before deleting author '{author.Name}'."
+            );
+        }
+
+        author.SetDefaultsForExtraProperties();// Ensure extra properties are set before deletion
+
         await _authorRepository.DeleteAsync(id);
 
-        await _backgroundJobManager.EnqueueAsync(
-            new EmailSendingArgs(
-                "phucanhbtt@gmail.com",
-                $"An author has been deleted by user {_currentUser.Name + _currentUser.SurName}.",
-                $"Author {author.Name} has just been updated."
-                )
-            );
+        await _notificationAppService.InsertNotificationAndSendEmailAsync(
+            _currentUser.Id.GetValueOrDefault(),
+            _currentUser.Email,
+            NotificationType.AuthorCRUD,
+            ObjectMapper.Map<Author, AuthorDto>(author),
+            "deleted");
+    }
+
+    public async Task<DateTime> GetMinDateTimeAsync()
+    {
+        return await _authorRepository.MinAsync(author => author.BirthDate);
+    }
+
+    private static string NormalizeSorting(string sorting)
+    {
+        if (sorting.IsNullOrEmpty())
+        {
+            return nameof(AuthorDto.Name); // Mặc định sắp xếp theo Name của BookDto
+        }
+
+        // Danh sách các thuộc tính hợp lệ của BookDto
+        var validSortProperties = new[]
+        {
+            nameof(AuthorDto.Name),
+            nameof(AuthorDto.BirthDate)
+        };
+
+        // Tách sorting thành tên thuộc tính và hướng sắp xếp (asc/desc)
+        var sortParts = sorting.Split(' ', StringSplitOptions.RemoveEmptyEntries);
+        var propertyName = sortParts[0];
+        var sortDirection = sortParts.Length > 1 ? sortParts[1].ToLower() : "asc";
+
+        // Kiểm tra xem propertyName có hợp lệ không
+        if (validSortProperties.Contains(propertyName, StringComparer.OrdinalIgnoreCase))
+        {
+            return sorting; // Trả về nguyên sorting nếu thuộc tính hợp lệ
+        }
+
+        // Nếu thuộc tính không hợp lệ, trả về mặc định
+        return nameof(AuthorDto.Name);
     }
 }
